Remove all translation from the back-end setting page
This feature, configuration for display card brand logo, has not yet
included the translation feature.

Although, using the translation function without specify the text domain
or specify the invalid text domain will not affect the system to stop
working but if using the translation function, the translation file (MO
file) should be available.