Correct english grammar of an error message
Correct the english grammar of an error message, when failed to create
Omise charge.

The error message will be appeared after created the Omise charge and
then retrieved the detail of Omise charge but the charge status is NOT
authorized (in case of manual capture) or NOT paid (in case of auto capture).