Get rid of 'omise-api-wrapper.php' class (use Omise-PHP instead).
Since we already used Omise-PHP in the plugin, manage 2 places of API connector would make it hard to maintain.
Also, we usually update Omise-PHP to support new Omise APIs. So, if Omise-WooCommerce needs to use new API, we don't need to update the connector in the plugin but just update Omise-PHP lib.